Changed default Wire (I2C) to use hardware I2C rather than bit banged software. The existing software Wire implementation has been retained as SoftWare.  Examples have also been updated to include Wire, HardWire and Softwire.
