+++ STM32F1/libraries/Wire/Wire.cpp
@@ -25,222 +25,79 @@
  *****************************************************************************/
 
 /**
- * @file Wire.cpp
+ * @file HardWire.cpp
  * @author Trystan Jones <crenn6977@gmail.com>
- * @brief Wire library, uses the WireBase to create the primary interface
- *        while keeping low level interactions invisible to the user.
+ * @brief Wire library, uses the hardware I2C available in the Maple to
+ *        interact with I2C slave devices.
  */
 
 /*
- * Library updated by crenn to follow new Wire system.
- * Code was derived from the original Wire for maple code by leaflabs and the
- * modifications by gke and ala42.
+ * Library created by crenn to use the new WireBase system and allow Arduino
+ * users easy interaction with the I2C Hardware in a familiar method.
  */
- /*
-  * Updated by Roger Clark. 20141111. Fixed issue when process() returned because of missing ACK (often caused by invalid device address being used), caused SCL to be left
-  * LOW so that in the next call to process() , the first clock pulse was not sent, because SCL was LOW when it should have been high.
-  */
- /*
-  * Updated by Brandon Green. 20172306. Implementing the repeated start functionality.
-  */
 
 #include "Wire.h"
 
-#define I2C_WRITE 0
-#define I2C_READ  1
-
-/* low level conventions:
- * - SDA/SCL idle high (expected high)
- * - always start with i2c_delay rather than end
- */
-
-void TwoWire::set_scl(bool state) {
-    I2C_DELAY(this->i2c_delay);
-
-	gpio_write_bit(sclDevice,sclBit, state);
-//    digitalWrite(this->scl_pin,state);
-    //Allow for clock stretching - dangerous currently
-    if (state == HIGH) {
-        while(digitalRead(this->scl_pin) == 0);
-    }
-}
-
-void TwoWire::set_sda(bool state) {
-	I2C_DELAY(this->i2c_delay);
-	gpio_write_bit(sdaDevice,sdaBit, state);
-    //digitalWrite(this->sda_pin, state);
-}
-
-void TwoWire::i2c_start() {
-    set_sda(LOW);
-    set_scl(LOW);
-}
-
-void TwoWire::i2c_stop() {
-    set_sda(LOW);
-    set_scl(HIGH);
-    set_sda(HIGH);
-}
-
-void TwoWire::i2c_repeated_start() {
-    set_sda(HIGH);
-    set_scl(HIGH);
-    set_sda(LOW);
-}
-
-bool TwoWire::i2c_get_ack() {
-    set_scl(LOW);
-    set_sda(HIGH);
-    set_scl(HIGH);
-
-    bool ret = !digitalRead(this->sda_pin);
-    set_scl(LOW);
-    return ret;
-}
-
-void TwoWire::i2c_send_ack() {
-    set_sda(LOW);
-    set_scl(HIGH);
-    set_scl(LOW);
-}
-
-void TwoWire::i2c_send_nack() {
-    set_sda(HIGH);
-    set_scl(HIGH);
-    set_scl(LOW);
-}
-
-uint8 TwoWire::i2c_shift_in() {
-    uint8 data = 0;
-    set_sda(HIGH);
-
-    int i;
-    for (i = 0; i < 8; i++) {
-        set_scl(HIGH);
-        data |= digitalRead(this->sda_pin) << (7-i);
-        set_scl(LOW);
-    }
-
-    return data;
-}
-
-void TwoWire::i2c_shift_out(uint8 val) {
-    int i;
-    for (i = 0; i < 8; i++) {
-        set_sda(!!(val & (1 << (7 - i)) ) );
-        set_scl(HIGH);
-        set_scl(LOW);
-    }
-}
-
-//process needs to be updated for repeated start.
-uint8 TwoWire::process(uint8 stop) {
-    itc_msg.xferred = 0;
-
-    uint8 sla_addr = (itc_msg.addr << 1);
-    if (itc_msg.flags == I2C_MSG_READ) {
-        sla_addr |= I2C_READ;
-    }
-    i2c_start();
-    // shift out the address we're transmitting to
-    i2c_shift_out(sla_addr);
-    if (!i2c_get_ack()) 
-	{
-		i2c_stop();// Roger Clark. 20141110 added to set clock high again, as it will be left in a low state otherwise
-        return ENACKADDR;
-    }
-    // Recieving
-    if (itc_msg.flags == I2C_MSG_READ) {
-        while (itc_msg.xferred < itc_msg.length) {
-            itc_msg.data[itc_msg.xferred++] = i2c_shift_in();
-            if (itc_msg.xferred < itc_msg.length) 
-			{
-                i2c_send_ack();
-            } 
-			else 
-			{
-                i2c_send_nack();
-            }
+uint8 HardWire::process(uint8 stop) {
+    int8 res = i2c_master_xfer(sel_hard, &itc_msg, 1, 0);
+    if (res == I2C_ERROR_PROTOCOL) {
+        if (sel_hard->error_flags & I2C_SR1_AF) { /* NACK */
+            res = (sel_hard->error_flags & I2C_SR1_ADDR ? ENACKADDR : 
+                                                          ENACKTRNS);
+        } else if (sel_hard->error_flags & I2C_SR1_OVR) { /* Over/Underrun */
+            res = EDATA;
+        } else { /* Bus or Arbitration error */
+            res = EOTHER;
         }
+        i2c_disable(sel_hard);
+        i2c_master_enable(sel_hard, (I2C_BUS_RESET | dev_flags));
     }
-    // Sending
-    else {
-        for (uint8 i = 0; i < itc_msg.length; i++) {
-            i2c_shift_out(itc_msg.data[i]);
-            if (!i2c_get_ack()) 
-			{
-				i2c_stop();// Roger Clark. 20141110 added to set clock high again, as it will be left in a low state otherwise
-                return ENACKTRNS;
-            }
-            itc_msg.xferred++;
-        }
-    }
-    if(stop == true)
-		i2c_stop();
-    else i2c_repeated_start();
-	
-    return SUCCESS;
+    return res;
 }
 
-// For compatibility with legacy code
-uint8 TwoWire::process(){
+uint8 HardWire::process(){
 	return process(true);
 }
 
-// TODO: Add in Error Handling if pins is out of range for other Maples
-// TODO: Make delays more capable
-TwoWire::TwoWire(uint8 scl, uint8 sda, uint8 delay) : i2c_delay(delay) {
-    this->scl_pin=scl;
-    this->sda_pin=sda;
+// TODO: Add in Error Handling if devsel is out of range for other Maples
+HardWire::HardWire(uint8 dev_sel, uint8 flags) {
+    if (dev_sel == 1) {
+        sel_hard = I2C1;
+    } else if (dev_sel == 2) {
+        sel_hard = I2C2;
+    } else {
+        ASSERT(1);
+    }
+    dev_flags = flags;
 }
 
-void TwoWire::begin(uint8 self_addr) {
-    tx_buf_idx = 0;
-    tx_buf_overflow = false;
-    rx_buf_idx = 0;
-    rx_buf_len = 0;
-    pinMode(this->scl_pin, OUTPUT_OPEN_DRAIN);
-    pinMode(this->sda_pin, OUTPUT_OPEN_DRAIN);
-	
-	sclDevice = PIN_MAP[this->scl_pin].gpio_device;
-	sclBit = PIN_MAP[this->scl_pin].gpio_bit;	
-	sdaDevice = PIN_MAP[this->sda_pin].gpio_device;
-	sdaBit = PIN_MAP[this->sda_pin].gpio_bit;		
-    set_scl(HIGH);
-    set_sda(HIGH);
+HardWire::~HardWire() {
+    i2c_disable(sel_hard);
+    sel_hard = 0;
 }
 
-void TwoWire::end()
-{
-	if (this->scl_pin)
-	{
-		pinMode(this->scl_pin, INPUT);
-	}
-	if (this->sda_pin)
-	{
-		pinMode(this->sda_pin, INPUT);
-	}
+void HardWire::begin(uint8 self_addr) {
+    i2c_master_enable(sel_hard, dev_flags);
 }
 
-void TwoWire::setClock(uint32_t frequencyHz)
+void HardWire::end() {
+    i2c_disable(sel_hard);
+    sel_hard = 0;
+}
+
+void HardWire::setClock(uint32_t frequencyHz)
 {
 	switch(frequencyHz)
 	{
 		case 400000:
-			i2c_delay = SOFT_FAST;
+			dev_flags |= I2C_FAST_MODE;// set FAST_MODE bit
 			break;
 		case 100000:
 		default:
-			i2c_delay = SOFT_STANDARD;
+			dev_flags &= ~I2C_FAST_MODE;// clear FAST_MODE bit
 			break;
 	}
+	
 }
 
-TwoWire::~TwoWire() {
-    this->scl_pin=0;
-    this->sda_pin=0;
-}
-
-// Declare the instance that the users of the library can use
-TwoWire Wire(SCL, SDA, SOFT_STANDARD);
-//TwoWire Wire(PB6, PB7, SOFT_STANDARD);
+HardWire Wire(1, I2C_FAST_MODE);;
