+++ STM32F1/libraries/Wire/examples/i2c_scanner_hardwire/i2c_scanner_hardwire.ino
@@ -26,7 +26,7 @@
 // Devices with higher bit address might not be seen properly.
 //
 
-#include <HardWire.h>
+#include <Wire.h>
 
 HardWire HWire(1, I2C_FAST_MODE); // I2c1
