+++ STM32F1/libraries/Wire/examples/i2c_scanner_softwire/i2c_scanner_softwire.ino
@@ -1,0 +1,54 @@
+// --------------------------------------
+// i2c_scanner
+//
+//
+
+#include <SoftWire.h>
+
+TwoWire SWire(PB6, PB7, SOFT_FAST);
+
+
+void setup() {
+  Serial.begin(115200);
+  SWire.begin();
+  Serial.println("\nSoftware I2C.. Scanner");
+}
+
+
+void loop() {
+  byte error, address;
+  int nDevices;
+
+  Serial.println("Scanning...");
+
+  nDevices = 0;
+  for(address = 1; address < 127; address++) {
+    // The i2c_scanner uses the return value of
+    // the Write.endTransmisstion to see if
+    // a device did acknowledge to the address.
+
+    SWire.beginTransmission(address);
+    error = SWire.endTransmission();
+    
+    if (error == 0) {
+      Serial.print("I2C device found at address 0x");
+      if (address < 16) 
+        Serial.print("0");
+      Serial.println(address, HEX);
+
+      nDevices++;
+    }
+    else if (error == 4) {
+      Serial.print("Unknown error at address 0x");
+      if (address < 16) 
+        Serial.print("0");
+      Serial.println(address, HEX);
+    }    
+  }
+  if (nDevices == 0)
+    Serial.println("No I2C devices found");
+  else
+    Serial.println("done");
+
+  delay(5000);           // wait 5 seconds for next scan
+}
